Move album refresh to scanner
